more images compressed and files renamed
