+++ src/assets/committee/Scott.jpg
